backend/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
websockets>=12.0
pydantic>=2.0.0

# Audio Processing
librosa>=0.9.1
numpy>=1.20.0
soundfile>=0.10.3
scipy>=1.12.0
pydub>=0.25.1

# Machine Learning
scikit-learn>=1.4.0

# Utilities
python-dotenv>=0.19.0
python-jose>=3.3.0
passlib>=1.7.4
aiofiles>=23.2.1
requests>=2.31.0

# Development
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

# Whisper
openai-whisper==20231117
torch>=2.0.0